refactor: 게임-경기일정-방송안내 리팩토링
