+++ src/components/Game/Schedule/Broadcast/BroadcastItem.tsx
@@ -1,8 +1,19 @@
+import { TBroadcastTag } from "@customTypes/game/schedule";
+import { broadcastItemList } from "@data/game/broadcast";
 import styled from "styled-components";
+import Tag from "./Tag";
 
-interface BroadcastItemProps {
-  tag: string;
-}
+const BroadcastListStyle = styled.li`
+  display: flex;
+  align-items: baseline;
+
+  & > div {
+    width: 100%;
+    display: flex;
+    flex-wrap: wrap;
+    gap: 5px;
+  }
+`;
 
 const BroadcastItemStyle = styled.p`
   font-size: 13px;
@@ -28,155 +39,20 @@ const BroadcastItemStyle = styled.p`
   }
 `;
 
-const BroadcastItem = ({ tag }: BroadcastItemProps) => {
-  const broadcastItemList = [
-    {
-      tv: [
-        {
-          abb: "K-2T",
-          name: "KBS 2TV",
-        },
-        {
-          abb: "M-T",
-          name: "MBC TV",
-        },
-        {
-          abb: "S-T",
-          name: "SBS TV",
-        },
-        {
-          abb: "KN-T",
-          name: "KBS N SPORTS",
-        },
-        {
-          abb: "MS-T",
-          name: "MBC SPORTS PLUS",
-        },
-        {
-          abb: "SS-T",
-          name: "SBS SPORTS",
-        },
-        {
-          abb: "SPOTV+",
-          name: "SPOTV PLUS",
-        },
-        {
-          abb: "SKY-T",
-          name: "SKY SPORTS",
-        },
-      ],
-      cmb: [
-        {
-          abb: "D-CMB",
-          name: "대전 CMB",
-        },
-        {
-          abb: "K-CMB",
-          name: "광주 CMB",
-        },
-      ],
-      iptv: [
-        {
-          abb: "SPO-T",
-          name: "SPOTV",
-        },
-        {
-          abb: "SPO-2T",
-          name: "SPOTV2",
-        },
-        {
-          abb: "IB-T",
-          name: "IB SPORTS",
-        },
-      ],
-      radio: [
-        {
-          abb: "K-2R",
-          name: "KBS 2라디오",
-        },
-        {
-          abb: "M-R",
-          name: "MBC 라디오",
-        },
-        {
-          abb: "S-R",
-          name: "SBS 라디오",
-        },
-        {
-          abb: "DK-R",
-          name: "대전 KBS 라디오",
-        },
-        {
-          abb: "TM-R",
-          name: "대구 MBC 라디오",
-        },
-        {
-          abb: "PM-R",
-          name: "부산 MBC 라디오",
-        },
-        {
-          abb: "DM-R",
-          name: "대전 MBC 라디오",
-        },
-        {
-          abb: "KM-R",
-          name: "광주 MBC 라디오",
-        },
-        {
-          abb: "GM-R",
-          name: "MBC 경남 라디오",
-        },
-        {
-          abb: "T-R",
-          name: "TBC 라디오",
-        },
-        {
-          abb: "TJ-R",
-          name: "TJB 라디오",
-        },
-        {
-          abb: "KNN-R",
-          name: "KNN 라디오",
-        },
-        {
-          abb: "KBC-R",
-          name: "KBC 라디오",
-        },
-      ],
-    },
-  ];
-
+const BroadcastItem = ({ tag }: { tag: TBroadcastTag }) => {
   return (
     <>
-      {broadcastItemList.map((item) =>
-        tag === "TV"
-          ? item.tv.map((el, index) => (
-              <BroadcastItemStyle key={index}>
-                {el.abb}
-                <span>{`(${el.name})`}</span>
-              </BroadcastItemStyle>
-            ))
-          : tag === "CMB"
-            ? item.cmb.map((el, index) => (
-                <BroadcastItemStyle key={index}>
-                  {el.abb}
-                  <span>{`(${el.name})`}</span>
-                </BroadcastItemStyle>
-              ))
-            : tag === "IPTV"
-              ? item.iptv.map((el, index) => (
-                  <BroadcastItemStyle key={index}>
-                    {el.abb}
-                    <span>{`(${el.name})`}</span>
-                  </BroadcastItemStyle>
-                ))
-              : item.radio.map((el, index) => (
-                  <BroadcastItemStyle key={index}>
-                    {el.abb}
-                    <span>{`(${el.name})`}</span>
-                  </BroadcastItemStyle>
-                ))
-      )}
+      <BroadcastListStyle>
+        <Tag tag={tag} />
+        <div>
+          {broadcastItemList[tag]?.map((item, index) => (
+            <BroadcastItemStyle key={index}>
+              {item.abb}
+              <span>{`(${item.name})`}</span>
+            </BroadcastItemStyle>
+          ))}
+        </div>
+      </BroadcastListStyle>
     </>
   );
 };
