+++ src/pages/game/Schedule.tsx
@@ -1,8 +1,6 @@
-import Broadcast from "@components/Game/Schedule/Broadcast";
-import Calendar from "@components/Game/Schedule/Calendar";
-import ScheduleList from "@components/Game/Schedule/ScheduleList";
-import { useEffect } from "react";
-import { useHomeStore } from "store/actions/useHomeStore";
+import Broadcast from "@components/Game/Schedule/Broadcast/Broadcast";
+import Calendar from "@components/Game/Schedule/Calendar/Calendar";
+import ScheduleList from "@components/Game/Schedule/RecentGame/ScheduleList";
 import styled from "styled-components";
 
 const ScheduleStyle = styled.section`
@@ -15,12 +13,6 @@ const ScheduleStyle = styled.section`
 `;
 
 const Schedule = () => {
-  const fetchRecentGames = useHomeStore((state) => state.fetchRecentGames);
-
-  useEffect(() => {
-    fetchRecentGames();
-  }, []);
-
   return (
     <ScheduleStyle>
       <ScheduleList />
