+++ src/types/game/watchPoint.d.ts
